requirements.txt
six
Faker
pillow
markdown
python-dateutil
pytz
DateTime
django
django-environ
django-filter
django-allauth
django-resized
python-decouple
django-crispy-forms
djangorestframework
dj-rest-auth
requests
crispy-bootstrap5
django-browser-reload
drf-yasg
django-phonenumber-field[phonenumbers]]
PyJWT
cryptography
paramiko
mailchimp_transactional
django-notifications-hq
fido2
django-jazzmin
autobahn
django_ckeditor_5
django-cities-light
django_otp
django-payments
django-storages[boto3]